refact(Update): Merge UpdateWithData into page/update.js
